style: Separating the models into base and system defined
