+++ openedx_tagging/core/tagging/models/__init__.py
@@ -1,0 +1,10 @@
+from .base import (
+    Tag,
+    Taxonomy,
+    ObjectTag,
+)
+
+from .system_defined import (
+    UserSystemDefinedTaxonomy,
+    LanguageTaxonomy
+)
